Web_App/requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
pymongo==4.6.1
cassandra-driver==3.29.0
kafka-python==2.0.2
python-socketio==5.10.0
eventlet==0.33.3
python-dotenv==1.0.0